# final exponentiation: select optimisation