@shramee 2024-11-26
Update regarding `bls12381` tests,
In tests I'm using `ecc.BN254.ScalarField()`, other tests seem to use bn254 field too so I'm guessing it's emulation of bls381 in bn254. But let me know if not and I'll fix it 🔧