@ivokub 2024-11-26
> Update regarding `bls12381` tests, In tests I'm using `ecc.BN254.ScalarField()`, other tests seem to use bn254 field too so I'm guessing it's emulation of bls381 in bn254. But let me know if not and I'll fix it 🔧

Indeed - gnark supports different curves for proving. In general it should work regardless of the curve used (BLS12-377, BN254 etc.), so choosing only one should work.