@ivokub 2024-11-26
> Thanks for the contribution! It looks good to me, but I'd like @yelhousni to confirm, in case we're moving towards direct extensions.

I pinged @yelhousni - lets see his opinion.