@shramee 2024-11-26
> ...in case we're moving towards direct extensions.
Direct extension for Schwartz Zippel/Fiat Shamir based ops?
https://github.com/shramee/cairo_pairing?tab=readme-ov-file#extension-field-operations