@ivokub 2024-11-26
> > ...in case we're moving towards direct extensions.
> 
> Direct extension for Schwartz Zippel/Fiat Shamir based ops like in here? https://github.com/shramee/cairo_pairing?tab=readme-ov-file#extension-field-operations

Exactly - we have already started some work here (for BW6-761): https://github.com/Consensys/gnark/pull/1312.